Add files via upload
Added CSS to button also an image for login page which will look nice on the login page.  To check the changes I have commented name and date that way it will easier for anyone to look at. Also, if we decide to remove the background image. let me know.